fix serializer
